requirements.txt
fpdf ~= 1.7.2
requests ~= 2.32.3
Pillow ~= 11.2.1